NOW, need to set spawnCondition better
