+++ core/src/main/java/com/libgdx/fallingblocks/Main.java
@@ -8,19 +8,12 @@ import com.libgdx.fallingblocks.screen.main.GameScreen;
 
 public class Main extends Game {
 
-    private InputListenerManager inputListenerManager;
 
     @Override
     public void create() {
-
-        this.inputListenerManager= new InputListenerManager();
-
-//        Logger.disableTag(Logger.Tags.GAME_OVER_STATE);
-//        Logger.disableTag(Logger.Tags.GAME_OVER_STATE);
-
         Logger.enableSingle(Logger.Tags.WAVE_PARSER);
         SpriteBatch spriteBatch = new SpriteBatch();
-        setScreen(new GameScreen(1, spriteBatch, inputListenerManager));
+        setScreen(new GameScreen(1, spriteBatch));
     }
 
     @Override
