+++ core/src/main/java/com/libgdx/fallingblocks/game/wave/WaveSettings.java
@@ -1,10 +1,10 @@
 package com.libgdx.fallingblocks.game.wave;
 
-import com.libgdx.fallingblocks.entity.enemy.difficulty.spawnrate.SpawnAbleConditions;
+import com.libgdx.fallingblocks.entity.enemy.difficulty.spawnrate.SpawnRateConditions;
 
 public class WaveSettings {
 
-    private SpawnAbleConditions spawnAbleConditions;
+    private SpawnRateConditions spawnRateConditions;
 
 
 }
